--- mainwindow.h
@@ -1,46 +1,0 @@
-#ifndef MAINWINDOW_H
-#define MAINWINDOW_H
-
-//For QClasses
-#include <QMainWindow>
-#include <QDialog>
-#include <QMessageBox>
-#include <QDebug>
-
-//For Background Image
-#include <QPixmap>
-
-#include "Login.h"
-#include "DisplayInfo.h"
-#include "SouvenirAndTrip.h"
-
-QT_BEGIN_NAMESPACE
-namespace Ui { class MainWindow; }
-QT_END_NAMESPACE
-
-class MainWindow : public QMainWindow
-{
-    Q_OBJECT
-
-public:
-    MainWindow(QWidget *parent = nullptr);
-    ~MainWindow();
-
-private slots:
-    ///This function will create a pointer to the login UI. The mainwindow will be
-    /// hidden before showing the login page
-	void on_Admin_Page_PushButton_clicked();
-
-    ///This function will create a pointer to the DisplayInfo UI. The mainwindow
-    /// will be hidden before showing the DisplayInfo page
-	void on_Display_Team_Stadium_Info_clicked();
-
-    ///This function will create a pointer to the Travel UI. The mainwindow will be
-    /// hidden before showing the Travel page
-	void on_Souvenir_And_Trip_PushButton_clicked();
-
-private:
-    Ui::MainWindow *ui;
-
-};
-#endif // MAINWINDOW_H
